adicionando testes da persistencia e do ranking

+++ TestFacade.py
@@ -9,10 +9,21 @@ from Facade import *
 class TestFacade(unittest.TestCase):
 
 	def setUp(self):
+		arquivoExiste= os.path.isfile("Ranking.dat")
+		if(arquivoExiste):
+			os.remove("Ranking.dat")
 		self.fac = Facade()
-		os.remove("Ranking.dat")
-	
-	#testes referentes a partidas e ao jogo
+
+	def ganharVida(self, n):
+		for i in range (0,n):
+			self.fac.ganharVida()
+
+	def perderVida(self, n):
+		for i in range (0,n):
+			self.fac.perderVida()
+
+
+		#testes referentes a partidas e ao jogo
 	def testeIniciarJogo(self):
 		self.fac.iniciarJogo()
 		self.assertEqual(True,self.fac.jogoComecou())
@@ -176,7 +187,6 @@ class TestFacade(unittest.TestCase):
 		except ExcecaoJogo as e : 
 			self.assertEqual(e.message, "Jogo não iniciado")
 			
-	#----------------------------* testesSistema *-----------------------------
 	#testes vidas (perder e ganhar)
 	def testeGanharVida(self):
 		self.fac.iniciarJogo()
@@ -283,40 +293,59 @@ class TestFacade(unittest.TestCase):
 		self.fac.iniciarJogo()
 		self.fac.iniciarPartida()
 		self.fac.criarNaveInimiga()
+		self.assertEqual(1, len(self.fac.getListaNaves()))
 		listaDeNaves = self.fac.getListaNaves()
 		self.fac.destruirNaveInimiga(listaDeNaves[0])
 		self.assertEqual(100, self.fac.getPontuacao())
+		self.assertEqual(0, len(self.fac.getListaNaves()))		
+		#varias naves
+		self.fac.criarNaveInimiga()
+		self.assertEqual(1, len(self.fac.getListaNaves()))
+		self.fac.criarNaveInimiga()
+		self.assertEqual(2, len(self.fac.getListaNaves()))
+		self.fac.criarNaveInimiga()
+		self.assertEqual(3, len(self.fac.getListaNaves()))
+		self.fac.criarNaveInimiga()
+		self.assertEqual(4, len(self.fac.getListaNaves()))
+		self.fac.destruirNaveInimiga(listaDeNaves[0])
+		self.fac.destruirNaveInimiga(listaDeNaves[0])
+		self.fac.destruirNaveInimiga(listaDeNaves[0])
+		self.fac.destruirNaveInimiga(listaDeNaves[0])
+		self.assertEqual(0, len(self.fac.getListaNaves()))
+		self.assertEqual(500, self.fac.getPontuacao())
 
 	#Testes do Ranking 	
-	def testeSairDoJogo(self):
+	def testeSairDaPartida(self):
 		self.fac.iniciarJogo()
 		self.fac.iniciarPartida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
+		self.fac.criarNaveInimiga()
+		self.fac.criarNaveInimiga()
+		self.assertEqual(2, len(self.fac.getListaNaves()))
+		self.fac.destruirNaveInimiga(self.fac.getListaNaves()[0])
+		self.fac.destruirNaveInimiga(self.fac.getListaNaves()[0])
+		self.assertEqual(0, len(self.fac.getListaNaves()))
+		self.ganharVida(6)
 		self.assertEqual(True, self.fac.sairDaPartida())
-		
-#========================| Testes para o 3o Estágio |========================#
 	
 	def testeCadastraNoRanking(self):
 		self.fac.iniciarJogo()
 		self.fac.iniciarPartida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
+		#ganhar pontos com vidas
+		self.ganharVida(8)
+		#ganhar pontos com alvos abatidos
+		self.fac.criarNaveInimiga()
+		self.fac.criarNaveInimiga()
+		self.fac.criarNaveInimiga()
+		self.assertEqual(3, len(self.fac.getListaNaves()))
+		self.fac.destruirNaveInimiga(self.fac.getListaNaves()[0])
+		self.fac.destruirNaveInimiga(self.fac.getListaNaves()[0])
+		self.fac.destruirNaveInimiga(self.fac.getListaNaves()[0])
+		self.assertEqual(0, len(self.fac.getListaNaves()))
 		self.assertTrue(self.fac.sairDaPartida())
 		self.fac.adicionarNomeNoRanking("Hermanoteu")	
 		self.assertEqual(1, len(self.fac.getRanking()))
 		self.assertEqual(self.fac.getRanking()[0][0],"Hermanoteu")
-		self.assertEqual(self.fac.getRanking()[0][1], 150)
+		self.assertEqual(self.fac.getRanking()[0][1], 450)
 	
 	def testeGetPontuacaoSemJogoIniciar(self):
 		try:
@@ -333,36 +362,109 @@ class TestFacade(unittest.TestCase):
 	def testeCadastraNoRankingDuasVezesComMesmaPontuacao(self):
 		self.fac.iniciarJogo()
 		self.fac.iniciarPartida()
+		#criando naves e ganhando vidas para conseguir pontos
 		self.fac.criarNaveInimiga()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
+		self.fac.criarNaveInimiga()
+		self.fac.criarNaveInimiga()
+		self.fac.criarNaveInimiga()
+		self.assertEqual(4, len(self.fac.getListaNaves()))
+		self.fac.destruirNaveInimiga(self.fac.getListaNaves()[0])
+		self.fac.destruirNaveInimiga(self.fac.getListaNaves()[0])
+		self.fac.destruirNaveInimiga(self.fac.getListaNaves()[0])
+		self.fac.destruirNaveInimiga(self.fac.getListaNaves()[0])
+		self.assertEqual(0, len(self.fac.getListaNaves()))
+		self.ganharVida(3)
+		self.assertEqual(0, len(self.fac.getRanking()))
 		self.assertTrue(self.fac.sairDaPartida())
-		self.fac.adicionarNomeNoRanking("Hermanoteu")	
-		
+		self.fac.adicionarNomeNoRanking("Hermanoteu")
+		self.assertEqual(1, len(self.fac.getRanking()))
+		self.fac.adicionarNomeNoRanking("Hermanoteu")
+		self.assertEqual(1, len(self.fac.getRanking()))
+	
+	def testeCadastraNoRankingEVerificarOrdem(self):
+		self.fac.iniciarJogo()
+		self.fac.iniciarPartida()
+		#criando naves e ganhando vidas para conseguir pontos
+		self.ganharVida(7)
+		self.assertEqual(100, self.fac.getPontuacao())
+		self.assertEqual(0, len(self.fac.getRanking()))
+		self.assertTrue(self.fac.sairDaPartida())
+		self.fac.adicionarNomeNoRanking("Hermanoteu")
+		self.fac.sairDoJogo
+		self.fac.iniciarPartida()
+		self.assertEqual(0, self.fac.getPontuacao())
+		self.ganharVida(6)
+		self.assertEqual(50, self.fac.getPontuacao())
+		self.assertEqual(1, len(self.fac.getRanking()))
+		self.assertTrue(self.fac.sairDaPartida())
+		self.fac.adicionarNomeNoRanking("raquel")
+		self.fac.iniciarPartida()
+		self.ganharVida(8)		
+		self.assertEqual(150, self.fac.getPontuacao())
+		self.assertEqual(2, len(self.fac.getRanking()))
+		self.assertTrue(self.fac.sairDaPartida())
+		self.fac.adicionarNomeNoRanking("Truta")
+		self.assertEqual(3, len(self.fac.getRanking()))
+		self.fac.iniciarPartida()
+		self.ganharVida(8)
+		self.assertEqual(150, self.fac.getPontuacao())
+		self.assertEqual(3, len(self.fac.getRanking()))
+		self.assertTrue(self.fac.sairDaPartida())
+		self.fac.adicionarNomeNoRanking("Feliphe")
+		self.assertEqual(4, len(self.fac.getRanking()))
+		self.assertEqual(self.fac.getRanking()[0][0],"Feliphe")
+		self.assertEqual(self.fac.getRanking()[0][1], 150)
+		self.assertEqual(self.fac.getRanking()[1][0],"Truta")
+		self.assertEqual(self.fac.getRanking()[1][1], 150)
+		self.assertEqual(self.fac.getRanking()[2][0],"Hermanoteu")
+		self.assertEqual(self.fac.getRanking()[2][1], 100)
+		self.assertEqual(self.fac.getRanking()[3][0],"raquel")
+		self.assertEqual(self.fac.getRanking()[3][1], 50)
+	
+	def testeGetNivel(self):
+		self.fac.iniciarJogo()
+		self.fac.iniciarPartida()
+		self.assertEqual(1, self.fac.getNivel())
+		self.assertTrue(self.fac.sairDaPartida())
+
+	def testeGetNivelAntesDoJogoIniciar(self):
+		try:
+			unittest.TestCase.assertRaises(self.fac.getNivel())
+		except ExcecaoJogo as e:
+			self.assertEqual(e.message, "Jogo não iniciado")
+	
+	def testeGetNivelAntesDaPartidaIniciar(self):
+		self.fac.iniciarJogo()
+		try:
+			unittest.TestCase.assertRaises(self.fac.getNivel())
+		except ExcecaoJogo as e:
+			self.assertEqual(e.message, "Partida não iniciada")
+	
+	def testeTerminarPartidaQuandoZerarVidasENaoAdicionarNoRankingPorNaoTerPontos(self):
+		self.fac.iniciarJogo()
+		self.fac.iniciarPartida()
+		self.perderVida(2)
+		self.assertTrue(self.fac.perderVida())
 		try:
 			unittest.TestCase.assertRaises(self.fac.adicionarNomeNoRanking("Hermanoteu"))
 		except ExcecaoJogo as e:
-			self.assertEqual(e.message, "Voce ja adcionou seu nome ao ranking")
-		self.assertEqual(1, len(self.fac.getRanking()))
-	
+			self.assertEqual(e.message, "Você precisa ter pontuação maior que zero")
 
-	def testePassarDeFase700Pontos(self):
+	def testeTerminarPartidaQuandoZerarVidasEAdicionarNoRanking(self):
 		self.fac.iniciarJogo()
 		self.fac.iniciarPartida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
+		self.ganharVida(20)
+		self.perderVida(7)
+		self.assertTrue(self.fac.perderVida())
+		self.fac.adicionarNomeNoRanking("Hermanoteu")
+		self.assertEqual(1, len(self.fac.getRanking()))
+		self.assertEqual(self.fac.getRanking()[0][0],"Hermanoteu")
+		self.assertEqual(self.fac.getRanking()[0][1], 750)
+
+	def testePassarDeNivelEm700Pontos(self):
+		self.fac.iniciarJogo()
+		self.fac.iniciarPartida()
+		self.ganharVida(8)
 		self.fac.criarNaveInimiga()
 		self.fac.criarNaveInimiga()
 		self.fac.criarNaveInimiga()
@@ -374,33 +476,148 @@ class TestFacade(unittest.TestCase):
 		self.fac.destruirNaveInimiga(listaDeNaves[0])
 		self.fac.destruirNaveInimiga(listaDeNaves[0])
 		self.fac.destruirNaveInimiga(listaDeNaves[0])
+		self.assertEqual(650, self.fac.getPontuacao())
+		#confere se ainda esta no nivel 1
+		self.assertEqual(1, self.fac.getNivel())
 		self.fac.ganharVida()
-		
-		self.assertTrue(self.fac.passouDeNivel())
+		#com a ultima vida ganha, passa para o nivel 2 com 700 pontos
 		self.assertEqual(700, self.fac.getPontuacao())
-		self.assertEqual(0, len(self.fac.getListaNaves()))
-	
-	#testes relacionados a camada de persistencia  (TODO)
-	def testeGravarRankingEmArquivo(self):
+		self.assertEqual(2, self.fac.getNivel())
+
+	def testePassarDeNivelAteFimDoJogo(self):
 		self.fac.iniciarJogo()
 		self.fac.iniciarPartida()
+		self.ganharVida(8)
+		self.fac.criarNaveInimiga()
+		self.fac.criarNaveInimiga()
+		self.fac.criarNaveInimiga()
+		self.fac.criarNaveInimiga()
+		self.fac.criarNaveInimiga()
+		listaDeNaves = self.fac.getListaNaves()
+		self.fac.destruirNaveInimiga(listaDeNaves[0])
+		self.fac.destruirNaveInimiga(listaDeNaves[0])
+		self.fac.destruirNaveInimiga(listaDeNaves[0])
+		self.fac.destruirNaveInimiga(listaDeNaves[0])
+		self.fac.destruirNaveInimiga(listaDeNaves[0])
+		self.assertEqual(650, self.fac.getPontuacao())
+		self.assertEqual(1, self.fac.getNivel())
 		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		self.fac.ganharVida()
-		
-		self.assertTrue(self.fac.sairDaPartida())
-		
+		#com a ultima vida ganha, passa para o nivel 2 com 700 pontos
+		self.assertEqual(700, self.fac.getPontuacao())
+		self.assertEqual(2, self.fac.getNivel())
+		self.ganharVida(14)
+		self.assertEqual(1400, self.fac.getPontuacao())
+		self.assertEqual(3, self.fac.getNivel())
+		self.ganharVida(14)
+		self.assertEqual(2100, self.fac.getPontuacao())
+		self.assertEqual(4, self.fac.getNivel())		
+		self.ganharVida(13)
+		self.assertFalse(self.fac.ganharVida())
+		self.assertEqual(2800, self.fac.getPontuacao())
+		self.assertEqual(5, self.fac.getNivel())
+		self.ganharVida(13)
+		self.assertTrue(self.fac.ganharVida())
 		self.fac.adicionarNomeNoRanking("Hermanoteu")
-		
-		#self.fac = Facade()
+
+	def testeGanharOuPerderVidaDepoisDeZerar(self):
+		self.fac.iniciarJogo()
+		self.fac.iniciarPartida()
+		self.ganharVida(18)
+		self.assertEqual(650, self.fac.getPontuacao())
+		self.assertEqual(1, self.fac.getNivel())
+		self.fac.ganharVida()
+		self.assertEqual(700, self.fac.getPontuacao())
+		self.assertEqual(2, self.fac.getNivel())
+		self.ganharVida(14)
+		self.assertEqual(1400, self.fac.getPontuacao())
+		self.assertEqual(3, self.fac.getNivel())
+		self.ganharVida(14)
+		self.assertEqual(2100, self.fac.getPontuacao())
+		self.assertEqual(4, self.fac.getNivel())		
+		self.ganharVida(13)
+		self.assertFalse(self.fac.ganharVida())
+		self.assertEqual(2800, self.fac.getPontuacao())
+		self.assertEqual(5, self.fac.getNivel())
+		self.ganharVida(13)
+		self.assertTrue(self.fac.ganharVida())
+		self.fac.adicionarNomeNoRanking("Hermanoteu")
+		#teste ganhar vida
+		try: 
+			unittest.TestCase.assertRaises(ExcecaoJogo, self.fac.ganharVida())
+		except ExcecaoJogo as e : 
+			self.assertEqual(e.message, "Partida não iniciada")
+		#teste Perder Vida			
+		try: 
+			unittest.TestCase.assertRaises(ExcecaoJogo, self.fac.perderVida())
+		except ExcecaoJogo as e : 
+			self.assertEqual(e.message, "Partida não iniciada")
+
+	def testeGravarRankingEmArquivoEVerificarSeDadosVoltamOrdenados(self):
+		self.fac.iniciarJogo()
+		self.fac.iniciarPartida()
+		self.ganharVida(8)
+		self.assertTrue(self.fac.sairDaPartida())
+		self.fac.adicionarNomeNoRanking("Hermanoteu")
 		self.assertEqual(1, len(self.fac.getRanking()))
 		self.assertEqual(self.fac.getRanking()[0][0],"Hermanoteu")
 		self.assertEqual(self.fac.getRanking()[0][1], 150)
-		
+		#verificando se hermanoteu foi recuperado com sucesso 
+		self.fac = Facade()	
+		self.assertEqual(1, len(self.fac.getRanking()))
+		self.assertEqual(self.fac.getRanking()[0][0],"Hermanoteu")
+		self.assertEqual(self.fac.getRanking()[0][1], 150)
+		self.fac.iniciarJogo()
+		self.fac.iniciarPartida()
+		self.ganharVida(8)
+		self.assertTrue(self.fac.sairDaPartida())
+		self.fac.adicionarNomeNoRanking("Manolo")
+		self.assertEqual(2, len(self.fac.getRanking()))
+		self.assertEqual(self.fac.getRanking()[0][0],"Hermanoteu")
+		self.assertEqual(self.fac.getRanking()[0][1], 150)
+		self.assertEqual(self.fac.getRanking()[1][0],"Manolo")
+		self.assertEqual(self.fac.getRanking()[1][1], 150)
+		#Verificando se Manolo e hermanoteu foram recuperados com sucesso e ornados alfabeticamente 
+		self.fac = Facade()
+		self.assertEqual(2, len(self.fac.getRanking()))
+		self.assertEqual(self.fac.getRanking()[0][0],"Hermanoteu")
+		self.assertEqual(self.fac.getRanking()[0][1], 150)
+		self.assertEqual(self.fac.getRanking()[1][0],"Manolo")
+		self.assertEqual(self.fac.getRanking()[1][1], 150)
+		self.fac.iniciarJogo()
+		self.fac.iniciarPartida()
+		self.ganharVida(9)
+		self.assertTrue(self.fac.sairDaPartida())
+		self.fac.adicionarNomeNoRanking("Truta")
+		self.assertEqual(3, len(self.fac.getRanking()))
+		self.assertEqual(self.fac.getRanking()[0][0],"Truta")
+		self.assertEqual(self.fac.getRanking()[0][1], 200)
+		self.assertEqual(self.fac.getRanking()[1][0],"Hermanoteu")
+		self.assertEqual(self.fac.getRanking()[1][1], 150)
+		self.assertEqual(self.fac.getRanking()[2][0],"Manolo")
+		self.assertEqual(self.fac.getRanking()[2][1], 150)
+		#Verificando se Hermanoteu, Manolo e Truta foram recuperados com sucesso(primeiro em pontuacao e depois em ordem alfabetica)
+		self.fac = Facade()
+		self.assertEqual(3, len(self.fac.getRanking()))
+		self.assertEqual(self.fac.getRanking()[0][0],"Truta")
+		self.assertEqual(self.fac.getRanking()[0][1], 200)
+		self.assertEqual(self.fac.getRanking()[1][0],"Hermanoteu")
+		self.assertEqual(self.fac.getRanking()[1][1], 150)
+		self.assertEqual(self.fac.getRanking()[2][0],"Manolo")
+		self.assertEqual(self.fac.getRanking()[2][1], 150)		
+		self.fac.iniciarJogo()
+		self.fac.iniciarPartida()
+		self.ganharVida(7)
+		self.assertTrue(self.fac.sairDaPartida())
+		self.fac.adicionarNomeNoRanking("Micalateia")
+		self.assertEqual(4, len(self.fac.getRanking()))
+		self.fac = Facade()
+		self.assertEqual(self.fac.getRanking()[0][0],"Truta")
+		self.assertEqual(self.fac.getRanking()[0][1], 200)
+		self.assertEqual(self.fac.getRanking()[1][0],"Hermanoteu")
+		self.assertEqual(self.fac.getRanking()[1][1], 150)
+		self.assertEqual(self.fac.getRanking()[2][0],"Manolo")
+		self.assertEqual(self.fac.getRanking()[2][1], 150)
+		self.assertEqual(self.fac.getRanking()[3][0],"Micalateia")
+		self.assertEqual(self.fac.getRanking()[3][1], 100)
 
 unittest.main()
